backend/requirements.txt
# JIT Orchestrator Dependencies

# AGNO framework (install from local)
# pip install -e ./agno/libs/agno

# Core dependencies
python-dotenv>=1.0.0    # Load .env files
google-genai>=1.0.0     # Google Gemini SDK
openai>=1.0.0           # OpenAI SDK (for OpenRouter)
mcp>=1.0.0              # MCP client
pydantic>=2.0.0         # Data validation
tiktoken>=0.5.0         # Token counting for session memory

# Web interface
fastapi>=0.104.0        # Web framework
uvicorn[standard]>=0.24.0  # ASGI server
websockets>=12.0        # WebSocket support

# ============================================
# MCP Servers (installed via uvx/npx on demand)
# ============================================
# These are NOT pip dependencies - they run as separate processes
#
# npm (via npx):
#   - @modelcontextprotocol/server-brave-search
#   - @modelcontextprotocol/server-filesystem
#   - @modelcontextprotocol/server-github
#   - @h1deya/mcp-server-weather
#   - @notionhq/notion-mcp-server
#   - geoapify-mcp-server
#
# Python (via uvx):
#   - openweather-mcp
#   - mcp-telegram
#   - workspace-mcp
#   - spotify-mcp
